Map Item is now working (need some more tweak)
added IsItemInHand() method to PlayerManager for checking if certain types of ItemCard is in the player's hand.

We then use that method to check in the line 149 of DeckManager when the player is drawing a card.

Changed Drawn interface from void to IEnumerator to use yield return when event is happening, we should wait for n event to finish first to start the next event. This problem is discovered when player has a map and drawn 2 event cards back-to-back.